+++ packages/helpers/src/types.ts
@@ -1,0 +1,1 @@
+export type PropertyName<TTarget> = TTarget extends Map<string, any> ? string : keyof TTarget & string;