+++ packages/views/src/bindingComponent.tsx
@@ -1,4 +1,4 @@
-import { ensureObservableProperty } from "@frui.ts/helpers";
+import { ensureObservableProperty, PropertyName } from "@frui.ts/helpers";
 import { action, get, isObservable, isObservableMap, isObservableProp } from "mobx";
 import * as React from "react";
 
@@ -14,7 +14,7 @@ export interface IBindingProps<TTarget> {
   target?: TTarget;
 
   /** Name of the bound property on the target entity */
-  property?: keyof TTarget & string;
+  property?: PropertyName<TTarget>;
 
   /**
    * Event handler called when the value of the control is changed
@@ -23,7 +23,7 @@ export interface IBindingProps<TTarget> {
    * @param property  Name of the bound property on the target entity
    * @param target  The target entity for the binding
    */
-  onValueChanged?: (value: any, property: keyof TTarget & string, target: TTarget) => void;
+  onValueChanged?: (value: any, property: PropertyName<TTarget>, target: TTarget) => void;
 }
 
 export type ExcludeBindingProps<T> = Omit<T, keyof IBindingProps<any>>;
@@ -71,9 +71,13 @@ export abstract class BindingComponent<TTarget, TProps extends IBindingProps<TTa
       throw new Error("'property' prop has not been set");
     }
 
+    if (isObservableMap(target)) {
+      return target.get(property);
+    }
+
     // TODO enable disable of automatically creating observable
-    if (!isObservable(target) || (!isObservableMap(target) && !isObservableProp(target, property))) {
-      ensureObservableProperty(target, property, target[property]);
+    if (!isObservable(target) || !isObservableProp(target, property)) {
+      ensureObservableProperty(target, property, target[property as keyof TTarget]);
     }
 
     return get(target, property);